// src/app/master/model/assetsmaster.ts
export class AssetmasterMapping {
    assets: string = "";
    assetstype: string = "";
    prefixcode: string = "";
    status: boolean = false;
 
  
 
}

export class Assetsselect {
    assets: string = "";
    assetscode: number = 0;
}

export class Assetstypeselect {
    assetstype: string = "";
    assetstypecode: number = 0;
}

export class Prefixcodeselect {
    prefixcodename: string = "";
    prefixcode: number = 0;
}